refactor: rename and refactor supported platform components
