+++ components/download/SupportedPlatformItem.tsx
