+++ components/sections/download/supported-platform.tsx
@@ -2,9 +2,9 @@
 
 import { motion } from "motion/react";
 import { useInView } from "react-intersection-observer";
-import HeaderSection from "@/components/header-section";
-import { SupportedPlatformItem } from "@/components/supported-platform/supported-platform-item";
-import { supportedPlatforms } from "@/constants/supported-platforms";
+import SectionHeader from "@/components/SectionHeader";
+import { SupportedPlatformItem } from "@/components/download/SupportedPlatformItem";
+import { supportedPlatformsData } from "@/constants/supportedPlatformsData";
 
 export const SupportedPlatform = () => {
   const { ref, inView } = useInView({
@@ -20,10 +20,10 @@ export const SupportedPlatform = () => {
         transition={{ duration: 0.6 }}
         className="px-5 lg:px-40"
       >
-        <HeaderSection section="downloadSupportedPlatform" />
+        <SectionHeader section="downloadSupportedPlatform" />
 
         <div className="mt-10 flex flex-row flex-wrap justify-center gap-4">
-          {supportedPlatforms.map((platform, index) => (
+          {supportedPlatformsData.map((platform, index) => (
             <motion.div
               key={platform.label}
               initial={{ opacity: 0, y: 20 }}
